requirements.txt
Flask==2.3.3
pymongo==4.6.1
pybreaker==1.0.0
pika==1.3.2
requests==2.28.1
python-dotenv==1.0.0